fixed api annotations
