+++ app/src/main/java/com/decagon/myrootapp/constants/URLEndpoints.kt
@@ -9,7 +9,7 @@ interface URLEndpoints {
         const val RESET_PASSWORD_ENDPOINT = "reset-password"
         const val TREE_ENDPOINT = "tree"
         const val SINGLE_TREE_ENDPOINT = "tree/{id}"
-        const val USER_TREE_ENDPOINT = "tree/{user}/tree"
+        const val USER_TREE_ENDPOINT = "tree/user/tree"
         const val PAYMENT_ENDPOINT = "payment"
         const val PAYMENT_UPDATE_ENDPOINT = "payment/{txref}"
         const val PAYMENT_UPDATE_STATUS_ENDPOINT = "payment/{txref}/status"
