+++ app/src/main/java/com/decagon/myrootapp/data/api/LoginAPI.kt
@@ -1,35 +1,33 @@
-package com.decagon.myrootapp.data.repositories
+package com.decagon.myrootapp.data.api
 
 import com.decagon.myrootapp.constants.URLConstants
 import com.decagon.myrootapp.constants.URLEndpoints
-import com.decagon.myrootapp.data.models.user.UserBody
-import com.decagon.myrootapp.data.models.user.UserResponse
+import com.decagon.myrootapp.data.models.login.LoginBody
+import com.decagon.myrootapp.data.models.login.LoginResponse
 import com.jakewharton.retrofit2.adapter.kotlin.coroutines.CoroutineCallAdapterFactory
 import kotlinx.coroutines.Deferred
 import okhttp3.OkHttpClient
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
-import retrofit2.http.*
+import retrofit2.http.Body
+import retrofit2.http.POST
 
-interface UserAPI {
+interface LoginAPI {
 
-    @POST(URLEndpoints.USER_ENDPOINT)
-    fun addUserAsync(
-        @Body user: UserBody
-    ): Deferred<UserResponse>
-
+    @POST(URLEndpoints.USER_LOGIN_ENDPOINT)
+    fun loginAsync(
+        @Body body: LoginBody
+    ): Deferred<LoginResponse>
 
     companion object{
-        fun invoke(): UserAPI{
-            val okHttpClient = OkHttpClient.Builder()
-                .build()
+        fun invoke(): LoginAPI{
+            val client = OkHttpClient.Builder().build()
             return Retrofit.Builder()
-                .baseUrl(URLConstants.BASE_URL)
                 .addCallAdapterFactory(CoroutineCallAdapterFactory())
                 .addConverterFactory(GsonConverterFactory.create())
-                .client(okHttpClient)
+                .baseUrl(URLConstants.BASE_URL)
                 .build()
-                .create(UserAPI::class.java)
+                .create(LoginAPI::class.java)
         }
     }
 }
